Connect Model and View
Connected the model and the view so that time left on the countdown page persists when the page is reloaded.

each time_manager page sets a specific machine

Restrictions:
I set a default user (User with id 1)
I made the user only be able to set the remaining time on their account to the machine.