Refactor query_time
Refactor query_time to separate the different functions in the thread. Noticed that the sync_switch_states takes a little bit of time and causes the thread to hang. We might have to switch this over to another thread.

Also have to figure out a way to update the time if the machine is closed via the update_expired_machines function

@Thanh2311 @digengill lmk what you think.

PR: #123